+++ test/spec/mrpreview.controller.ut.js
@@ -57,6 +57,7 @@
 
                 experience = {
                     id: 'foo',
+                    mode: 'light',
                     data: {
                         deck: [
                             {
@@ -121,7 +122,7 @@
                         });
 
                         it('should be "assets/apps/rumble/app/index.html?kCollateralUrl=../c6Content&kDebug=true&kDevMode=true"', function() {
-                            expect(controller().playerSrc).toBe('assets/apps/rumble/app/index.html?kCollateralUrl=' + encodeURIComponent('../c6Content') + '&kDebug=true&kDevMode=true');
+                            expect(controller().playerSrc).toBe('assets/apps/rumble/app/index.html?kCollateralUrl=' + encodeURIComponent('../c6Content') + '&kDebug=true&kDevMode=true&kDevice=desktop&kMode=full');
                         });
                     });
 
@@ -142,7 +143,22 @@
                         });
 
                         it('should be "/apps/rumble?kCollateralUrl=/collateral"', function() {
-                            expect(controller().playerSrc).toBe('/apps/rumble?kCollateralUrl=' + encodeURIComponent('/collateral'));
+                            expect(controller().playerSrc).toBe('/apps/rumble/?kCollateralUrl=' + encodeURIComponent('/collateral') + '&kDevice=desktop&kMode=full');
+                        });
+
+                        it('should pass the current mode and device', function() {
+                            spyOn(MiniReelService, 'convertForPlayer').and.returnValue(experience);
+                            $scope.$emit('mrPreview:initExperience', experience, session);
+
+                            expect(PreviewController.playerSrc).toBe('/apps/rumble/?kCollateralUrl=' + encodeURIComponent('/collateral') + '&kDevice=desktop&kMode=light');
+
+                            experience.mode = 'lightbox';
+                            $scope.$apply(function() {
+                                PreviewController.device = 'phone';
+                            });
+                            $scope.$emit('mrPreview:updateExperience', experience);
+
+                            expect(PreviewController.playerSrc).toBe('/apps/rumble/?kCollateralUrl=' + encodeURIComponent('/collateral') + '&kDevice=phone&kMode=lightbox');
                         });
                     });
 
@@ -163,7 +179,22 @@
                         });
 
                         it('should be "/apps/rumble?kCollateralUrl=/collateral"', function() {
-                            expect(controller().playerSrc).toBe('/apps/rumble?kCollateralUrl=' + encodeURIComponent('/collateral'));
+                            expect(controller().playerSrc).toBe('/apps/rumble/?kCollateralUrl=' + encodeURIComponent('/collateral') + '&kDevice=desktop&kMode=full');
+                        });
+
+                        it('should pass the current mode and device', function() {
+                            spyOn(MiniReelService, 'convertForPlayer').and.returnValue(experience);
+                            $scope.$emit('mrPreview:initExperience', experience, session);
+
+                            expect(PreviewController.playerSrc).toBe('/apps/rumble/?kCollateralUrl=' + encodeURIComponent('/collateral') + '&kDevice=desktop&kMode=light');
+
+                            experience.mode = 'lightbox';
+                            $scope.$apply(function() {
+                                PreviewController.device = 'phone';
+                            });
+                            $scope.$emit('mrPreview:updateExperience', experience);
+
+                            expect(PreviewController.playerSrc).toBe('/apps/rumble/?kCollateralUrl=' + encodeURIComponent('/collateral') + '&kDevice=phone&kMode=lightbox');
                         });
                     });
                 });
@@ -248,7 +279,6 @@
                         experience.data.deck.push(newCard);
 
                         $scope.$emit('mrPreview:updateExperience', experience);
-                        expect()
                         expect(MiniReelService.convertForPlayer).toHaveBeenCalled();
                     });
 
@@ -291,11 +321,12 @@
                 });
 
                 describe('mrPreview:updateMode', function() {
-                    it('should tell the player to reset', function() {
+                    it('should change the playerSrc, cause a refresh, and send an updated experience', function() {
                         var dataSentToPlayer,
                             emitCount = 0,
                             updatedExperience = {
                                 id: 'foo',
+                                mode: 'lightbox',
                                 data: {
                                     deck: [
                                         {
@@ -324,8 +355,10 @@
                             }
                         });
                         $scope.$emit('mrPreview:initExperience', experience, session);
+                        expect(PreviewController.playerSrc).toContain('kMode=light');
+                        
                         $scope.$emit('mrPreview:updateMode', updatedExperience);
-                        expect(session.ping.calls.argsFor(0)[0]).toBe('mrPreview:updateMode');
+                        expect(PreviewController.playerSrc).toContain('kMode=lightbox');
 
                         session.emit('handshake', {}, responseCallback);
 
@@ -345,6 +378,7 @@
                         beforeEach(function() {
                             updatedExperience = {
                                 id: 'foo',
+                                mode: 'light',
                                 data: {
                                     deck: [
                                         {
@@ -388,8 +422,8 @@
                             });
                         });
 
-                        it('should tell the player to reload', function() {
-                            expect(session.ping.calls.argsFor(2)[0]).toBe('mrPreview:updateMode');
+                        it('should cause the playerSrc to change', function() {
+                            expect(PreviewController.playerSrc).toContain('kDevice=phone');
                         });
 
                         it('should send an updated profile to the player after it reloads', function() {